requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep Learning
tensorflow>=2.8.0
torch>=1.12.0
keras>=2.8.0

# Time Series
statsmodels>=0.13.0
prophet>=1.1.0
tslearn>=0.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
streamlit>=1.10.0

# Data Processing
dask>=2022.1.0
pyarrow>=7.0.0
openpyxl>=3.0.0

# MLOps and Experimentation
mlflow>=1.25.0
wandb>=0.12.0
optuna>=2.10.0

# Feature Engineering
featuretools>=1.20.0
tsfresh>=0.19.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0
tqdm>=4.60.0
joblib>=1.1.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0
notebook>=6.4.0

# Testing and Quality
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Database and APIs
sqlalchemy>=1.4.0
requests>=2.27.0
fastapi>=0.75.0
uvicorn>=0.17.0

# Clustering and Advanced Analytics
hdbscan>=0.8.0
umap-learn>=0.5.0

# Financial/Cost Optimization
cvxpy>=1.2.0
ortools>=9.0.0
